requirements.txt
Django==1.7
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.18
celery==3.1.14
djangorestframework==2.4.2
elasticsearch==1.2.0
kombu==3.0.22
pytz==2014.7
urllib3==1.9
python-geohash==0.8.5
python-dateutil==2.2
